For each prefix, add a key <prefix> to RocksDB
This patch changes the disk format. It's backwards compatible (meaning,
you can upgrade), but it's not forwards compatible. Once you upgrade,
there is no going back to the old binary.

In this patch we add a key <prefix> to RocksDB for every prefix we
create. What this allows us to do is make sure that iterations on one
prefix don't end up on a different prefix. For example, consider this
scenario:

<prefix>a <prefix>b <prefix>c <prefix+1>Tombstone <prefix+1>Tombstone <prefix+1>Tombstone ...

Doing a Next() on <prefix>c will end up traversing all tombstones in
prefix+1, which is not really necessary. To avoid this, we add a key
<prefix+1>. With that barrier, Next() on a <prefix>c will end up on
<prefix+1> and avoid any tombstones from <prefix+1>